implementação do blockedActions no deleteService e editService

+++ src/services/components/serviceTable/Index.tsx
@@ -4,48 +4,62 @@ import { Service } from '../../../types/services'
 import { ServicesContext } from '../../../ServicesContext'
 import { changeFormState } from '../../../formFunctions/GenericForm'
 import { SERVER_URL } from '../../../App'
-import { DB_ERROR_TEXT, SERVER_ERROR_TEXT } from '../../Index'
+import { DB_ERROR_TEXT, SERVER_ERROR_TEXT, BLOCKED_ACTIONS_TEXT } from '../../Index'
 
 interface Props {
     setSearchKey: React.Dispatch<React.SetStateAction<string>>
     setEditForm: React.Dispatch<React.SetStateAction<boolean>>
+    blockedActions: boolean
+    setBlockedActions: React.Dispatch<React.SetStateAction<boolean>>
 }
 
-const Index = ({ setSearchKey, setEditForm }: Props) => {
+const Index = ({ setSearchKey, setEditForm, blockedActions, setBlockedActions }: Props) => {
 
     const [services, setServices] = useContext(ServicesContext)
 
     const deleteService = (targetService: Service) => {
-        changeFormState('', '')
-        const options = {
-            method: 'delete',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify(targetService)
+        if (!blockedActions) {
+            setBlockedActions(true)
+            changeFormState('', '')
+            const options = {
+                method: 'delete',
+                headers: { 'Content-Type': 'application/json' },
+                body: JSON.stringify(targetService)
+            }
+            setTimeout(() => {
+                fetch(`${SERVER_URL}/delete-service`, options)
+                .then(res => {
+                    switch (res.status) {
+                        case 204:
+                            const remainingServices = services.filter(service => {
+                                return service.name !== targetService.name
+                            })
+                            setServices(remainingServices)
+                            break
+                        case 503:
+                            alert(DB_ERROR_TEXT)
+                            break
+                    }
+                    setBlockedActions(false)
+                }).catch(() => {
+                    alert(SERVER_ERROR_TEXT)
+                    setBlockedActions(false)
+                })
+            }, 5000)
+        } else {
+            alert(BLOCKED_ACTIONS_TEXT)
         }
-        setTimeout(() => {
-            fetch(`${SERVER_URL}/delete-service`, options)
-            .then(res => {
-                switch (res.status) {
-                    case 204:
-                        const remainingServices = services.filter(service => {
-                            return service.name !== targetService.name
-                        })
-                        setServices(remainingServices)
-                        break
-                    case 503:
-                        alert(DB_ERROR_TEXT)
-                        break
-                }
-            }).catch(() => {
-                alert(SERVER_ERROR_TEXT)
-            })
-        }, 500)
     }
 
     const setEditValuesInTheForm = (name: string, value: number) => {
-        changeFormState(name, value.toString(), 'Editar Serviço')
-        setSearchKey(name)
-        setEditForm(true)
+        if (!blockedActions) {
+            setBlockedActions(true)
+            changeFormState(name, value.toString(), 'Editar Serviço')
+            setSearchKey(name)
+            setEditForm(true)
+        } else {
+            alert(BLOCKED_ACTIONS_TEXT)
+        }
     }
 
     return (
